+++ cli/src/lib.rs
@@ -22,10 +22,10 @@
 
 mod arg;
 mod cli;
-mod cmd_dashboard;
 mod cmd_event;
 mod cmd_generate_completion;
 mod cmd_todo;
+mod cmd_toplevel;
 mod cmd_tui;
 mod config;
 mod event_formatter;