+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - cli: Add `event delay` and `event reschedule` command to align with `todo` command group
+- cli: Add toplevel `delay` and `reschedule` command
 
 ## [v0.7.0] - 2025-09-08